clear up and fix pruning
dijkstra supposedly already handles that well, as the transitions pruned
here lead to redundant states on just the next step. I guess it makes
for less cost computations and priority queue updates.

(I had it disabled in my tests, and it had a corner case that does not
seem to change my solution.)